% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_update.R
\name{alpha_update}
\alias{alpha_update}
\title{Function to update tau_k given an Exp(lambda_2) prior}
\usage{
alpha_update(
  alpha_tau,
  tau_vec,
  lambda_2,
  mu_tau,
  sigma_tau,
  window,
  trunc = c(0, 100)
)
}
\arguments{
\item{alpha_tau}{numeric > 0, current value of alpha_tau shape parameter in tau_k prior}

\item{lambda_2}{numeric > 0, parameter to the exponential prior Exp(lambda_2)}

\item{window}{numeric > 0, parameter to set step size for tau update proposals from log-normal}

\item{trunc}{numeric > 0, truncation for regularization parameter tau_k}

\item{tau_k}{numeric > 0, current value of tau for Kth subject}
}
\description{
Function to update tau_k given an Exp(lambda_2) prior
}
