\name{bayesRCM-package}
\alias{bayesRCM-package}
\alias{bayesRCM}
\docType{package}
\title{
\packageTitle{bayesRCM}
}
\description{
\packageDescription{bayesRCM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayesRCM}
\packageIndices{bayesRCM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bayesRCM}

Maintainer: \packageMaintainer{bayesRCM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
