% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_glasso.R, R/rccmSummary.R
\name{bic_cal}
\alias{bic_cal}
\title{#BIC Calculations for tuning the lambda penalty in ind_graphs}
\usage{
bic_cal(x, Omegas, Gk_est = NULL)

bic_cal(x, Omegas, Gk_est = NULL)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{Omegas}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{Gk_est}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level networks.}
}
\value{
Numeric BIC value
}
\description{
This function calculates the BIC for the random covariance model (RCM)
}
\author{
Lin Zhang
}
