% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega0_update.R
\name{g0_log_density}
\alias{g0_log_density}
\title{Log density of omega_{0, (i, j)} posterior given bounds a, b}
\usage{
g0_log_density(kappa, nu, a, b, D)
}
\arguments{
\item{D}{}
}
\description{
Log density of omega_{0, (i, j)} posterior given bounds a, b
}
