% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_update.R
\name{graph_update}
\alias{graph_update}
\title{Function to update graph G, constraining the precision matrix Omega}
\usage{
graph_update(row_col, df, D, v, S, adj, omega, lambda_1)
}
\arguments{
\item{df}{Integer, degrees of freedom of the Gwishart distribution}

\item{D}{Matrix, location parameter of the Gwishart distribution}

\item{adj}{Matrix, current adjacency matrix/graph constraining precision matrix Omega}

\item{omega}{Matrix, current precision matrix Omega}
}
\description{
Function to update graph G, constraining the precision matrix Omega
}
