% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_update.R
\name{gwish_ij_update}
\alias{gwish_ij_update}
\title{Function to update an individual ij entry in G_k via method outline in Wang and Li (2011) "Efficient Gaussian Graphical Model Determination without Approximating Normalizing Constant of the G-Wishart distribution"}
\usage{
gwish_ij_update(b, D, omega, i, j, ij_cur)
}
\arguments{
\item{ij_cur}{}
}
\description{
Function to update an individual ij entry in G_k via method outline in Wang and Li (2011) "Efficient Gaussian Graphical Model Determination without Approximating Normalizing Constant of the G-Wishart distribution"
}
