% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_glasso.R
\name{ind_graphs}
\alias{ind_graphs}
\title{Independent glasso method for starting values}
\usage{
ind_graphs(x, lambda)
}
\arguments{
\item{lambda}{}
}
\description{
Independent glasso method for starting values
}
