% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda2_update.R
\name{lambda2_update}
\alias{lambda2_update}
\title{Function to update lambda_2 with Gamma prior}
\usage{
lambda2_update(
  lambda_2,
  alpha_tau,
  tau_vec,
  mu_tau,
  sigma_tau,
  window,
  trunc = c(0, 100)
)
}
\arguments{
\item{lambda_2}{numeric > 0, parameter to the exponential prior Exp(lambda_2)}

\item{alpha_tau}{numeric > 0, current value of alpha_tau shape parameter in tau_k prior}

\item{window}{numeric > 0, parameter to set step size for tau update proposals from log-normal}

\item{trunc}{numeric > 0, truncation for regularization parameter tau_k}

\item{tau_k}{numeric > 0, current value of tau for Kth subject}
}
\value{
Returns list of MH-step updated (or not) lambda_2 scalar value and a logicial T/F whether the proposal was accepted
}
\description{
Function to update
}
\examples{
lambda2_update(lambda_2, alpha_tau, tau_vec, mu_tau, sigma_tau, window, trunc = c(0, 100))
}
