% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_glasso.R, R/rccmSummary.R
\name{mbic_cal}
\alias{mbic_cal}
\title{Modified BIC for RCM}
\usage{
mbic_cal(x, Omega0, Omegas, lambda2, G0_est = NULL, Gk_est = NULL)

mbic_cal(x, Omega0, Omegas, lambda2, G0_est = NULL, Gk_est = NULL)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{Omega0}{\eqn{p} x \eqn{p} group-level precision matrix estimate.}

\item{Omegas}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}

\item{lambda2}{Non-negative scalar. Induces similarity between subject-level matrices and group-level matrix.}

\item{G0_est}{\eqn{p} x \eqn{p} group-level network estimate.}

\item{Gk_est}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level networks.}
}
\value{
Numeric Modified BIC value

Numeric Modified BIC value
}
\description{
This function calculates the modified BIC for the random covariance model (RCM)

This function calculates the modified BIC for the random covariance model (RCM)
}
\author{
Lin Zhang
}
