% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randCov.R
\name{randCov}
\alias{randCov}
\title{Random Covariance Model}
\usage{
randCov(x, lambda1, lambda2, lambda3 = 0)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{lambda1}{Non-negative scalar. Induces sparsity in subject-level matrices.}

\item{lambda2}{Non-negative scalar. Induces similarity between subject-level matrices and group-level matrix.}

\item{lambda3}{Non-negative scalar. Induces sparsity in group-level matrix.}
}
\value{
A list of length 2 containing:
\enumerate{
\item Group-level precision matrix estimate (Omega0).
\item \eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} subject-level precision matrix estimates (Omegas).
}
}
\description{
This function implements the Random Covariance Model (RCM) for joint estimation of
multiple sparse precision matrices. Optimization is conducted using block
coordinate descent.
}
\references{
Zhang, Lin, Andrew DiLernia, Karina Quevedo, Jazmin Camchong, Kelvin Lim, and Wei Pan.
"A Random Covariance Model for Bi-level Graphical Modeling with Application to Resting-state FMRI Data." 2019. https://arxiv.org/pdf/1910.00103.pdf
}
\author{
Lin Zhang
}
