% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm.R
\name{rcm}
\alias{rcm}
\title{Bayesian robust Random Covariance Model (RCM) for heirarchical graphical modeling}
\usage{
rcm(
  y = data_list,
  priors = NULL,
  n_samples = 100,
  n_burn = 10,
  n_cores = 4,
  n_updates = 2
)
}
\arguments{
\item{priors}{List of prior distribution specifications, defaults to -}

\item{n_samples}{Integer, number of posterior samples to obtain}

\item{n_burn}{Integer, number of posterior samples to burn}

\item{n_cores}{Integer, number of cores to be run in parallel - if empty/NULL will run sequentially}

\item{data_list}{List of $K$ subject's residual BOLD time series array (array of V volumes of matrices in time)}
}
\description{
Bayesian robust Random Covariance Model (RCM) for heirarchical graphical modeling
}
\examples{
rcm::example_data_list
rcm(example_data_list)
}
