% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate bayesRCM data with underlying G-Wishart data generating mechanism}
\usage{
sim_data(
  subjects = 20,
  volumes = 200,
  rois = 10,
  prop_true_con = 1/5,
  trunc = c(0, 100),
  alpha_tau = 50,
  lambda_2 = 2/5,
  n_flip = 1,
  seed_0 = 4,
  seed_k = 4,
  write = FALSE
)
}
\arguments{
\item{subjects}{K number of subjects}

\item{volumes}{V number of samples/fMRI observations (length of time-series)}

\item{rois}{P network dimension}

\item{prop_true_con}{Proportion of true connections}

\item{seed_0}{Reproducible seed for the simulation of G_0/Omega_0}

\item{seed_k}{Reproducible seed for the K simulations of subject-specific G_k/Omega_k}

\item{save}{Logical, if TRUE write save}
}
\value{
data_list
}
\description{
Simulate bayesRCM data with underlying G-Wishart data generating mechanism
}
