% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randCov.R
\name{spcov_bcd}
\alias{spcov_bcd}
\title{Covariance Graphical Lasso}
\usage{
spcov_bcd(samp_cov, rho, initial = NULL)
}
\arguments{
\item{samp_cov}{\eqn{p} x \eqn{p} sample covariance matrix.}

\item{rho}{Non-negative scalar. Induces sparsity in covariance matrix.}

\item{initial}{Initial value for covariance matrix.}
}
\value{
\eqn{p} x \eqn{p} sparse covariance matrix estimate.
}
\description{
This function implements the Random Covariance Model (RCM) for joint estimation of
multiple sparse precision matrices. Optimization is conducted using block
coordinate descent.
}
\references{
Wang, Hao.
"Two New Algorithms for Solving Covariance Graphical Lasso Based on Coordinate Descent and ECM." 2012. https://arxiv.org/pdf/1205.4120.pdf
}
