% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_update.R
\name{tau_update}
\alias{tau_update}
\title{Function to update tau_k given an Exp(lambda_2) prior}
\usage{
tau_update(
  tau_k,
  omega_k,
  sigma_0,
  alpha_tau,
  lambda_2,
  window,
  trunc = c(0, 100)
)
}
\arguments{
\item{tau_k}{numeric > 0, current value of tau for Kth subject}

\item{omega_k}{matrix, current subject specific precision matrix}

\item{sigma_0}{matrix, current group-level precision matrix}

\item{lambda_2}{#numeric > 0, parameter to the exponential prior Exp(lambda_2)}

\item{window}{#numeric > 0, parameter to set step size for tau update proposals from log-normal}
}
\description{
Function to update tau_k given an Exp(lambda_2) prior
}
